+++ src/main/java/pl/vabank/game/access/UserRepository.java
@@ -9,7 +9,7 @@ import pl.vabank.game.data.UserData;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface UserRepository extends JpaRepository<UserData, Integer> {
+public interface UserRepository extends JpaRepository<UserData, Long> {
     @Query("SELECT u FROM UserData u WHERE u.email = ?1")
     public UserData findByEmail(String email);
 }
